Rewrite the logging part to use SQLite instead of Git
Without periodic git-gc, the git backend is not very space-efficient.
However, git-gc becomes very expensive (both in time, CPU, and memory)
when there is a lot of data accumulated, eventually leading to failures.

This has now been replaced with a custom SQLite-backed system, where
only changes in user data are recorded.  Hopefully this will more
efficient.

This change introduced substantial changes, so the old data will not be
automatically migrated.